Split emit file-read host effect fixtures
