+++ docs/COMPLETION_AUDIT.md
@@ -333,7 +333,7 @@ and do not assume Phase 4 is ready without evidence.
   module.
 - Emit-command build graph host-effect integration tests now split declared
   file-read fallback accept/reject coverage into
-  `tests/integration/cli_build/emit_direct_host_effects/file_reads.rs`,
+  `tests/integration/cli_build/emit_direct_host_effects/file_reads/`,
   leaving env-effect ordering and unrelated test skip coverage in the parent
   module.
 - Build graph unit tests now split target metadata validation and gated
